Unit Test for Convert-ComputerDescription
Added a unit test for Convert-ComputerDescription.  I probably need
to refactor the prase description loop to make have it exclude trying
to match stuff that it has already matched instead of just going down
a list.

I need parameters in their that they can pass regex that will be used
for checking for AssetTags because everyone probably uses different
tags.

Also probably want an parameter to pass a description test and have
it be parsed which probably means creating a new test-function for
that as well as moving the parsing logic to it's own function.